user/requirements.txt
asgiref==3.4.1
daphne==3.0.2
Django>=3.2,<4.2.10
uvicorn
django-sslserver
requests_oauthlib
